Add useState
This is a higher level abstraction around the renderer which allows
us to automagically update the page when updating a value. The state
persists between navigations, so counters won't reset to zero
when navigating between routes.

The framework is still 69 lines of code of course, because who would
ever need more than 69 lines of code!?